Enforce egress allowlist in the control plane
