+++ data-plane/Cargo.toml
@@ -22,7 +22,6 @@ futures = "0.3.21"
 dns-message-parser = { version = "0.6.0", optional = true }
 tokio-rustls = { version = "0.23.4", features = ["dangerous_configuration"] }
 tokio-vsock = { version = "0.3.2", optional = true }
-tls-parser = { version = "0.11.0", optional = true }
 shared = { path = "../shared", default-features = false }
 dashmap = { version = "5.4.0", optional = true }
 serde = { version = "1.0.142", features = ["derive"] }
@@ -46,6 +45,6 @@ tokio-test = "0.4.2"
 [features]
 default = ["tls_termination"]
 tls_termination = ["dep:nom"]
-network_egress = ["dep:dns-message-parser", "dep:tls-parser", "dep:dashmap"]
+network_egress = ["dep:dns-message-parser", "dep:dashmap"]
 enclave = ["dep:tokio-vsock", "shared/enclave"]
 not_enclave = []
