+++ data-plane/src/configuration.rs
@@ -28,81 +28,3 @@ pub fn get_egress_ports() -> Vec<u16> {
         })
         .collect()
 }
-
-pub fn get_egress_allow_list() -> EgressDomains {
-    let domain_str = std::env::var("EV_EGRESS_ALLOW_LIST").unwrap_or("".to_string());
-    let (wildcard, exact): (Vec<String>, Vec<String>) = domain_str
-        .split(',')
-        .map(|domain| domain.to_string())
-        .partition(|domain| domain.starts_with("*."));
-    let wildcard_stripped = wildcard
-        .iter()
-        .filter_map(|wc| wc.strip_prefix('*').map(|domain| domain.to_string()))
-        .collect();
-    EgressDomains {
-        wildcard: wildcard_stripped,
-        exact: exact.clone(),
-        allow_all: exact == vec![""] || exact.contains(&"*".to_string()),
-    }
-}
-
-#[derive(Clone, PartialEq, Debug)]
-pub struct EgressDomains {
-    pub wildcard: Vec<String>,
-    pub exact: Vec<String>,
-    pub allow_all: bool,
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::configuration::{get_egress_allow_list, EgressDomains};
-
-    #[test]
-    fn test_sequentially() {
-        test_valid_all_domains();
-        test_wildcard_and_exact();
-        test_backwards_compat();
-    }
-
-    fn test_valid_all_domains() {
-        std::env::set_var("EV_EGRESS_ALLOW_LIST", "*");
-        let egress = get_egress_allow_list();
-        assert_eq!(
-            egress,
-            EgressDomains {
-                exact: vec!["*".to_string()],
-                wildcard: vec![],
-                allow_all: true
-            }
-        );
-        std::env::remove_var("EV_EGRESS_ALLOW_LIST");
-    }
-
-    fn test_wildcard_and_exact() {
-        std::env::set_var("EV_EGRESS_ALLOW_LIST", "*.evervault.com,google.com");
-        let egress = get_egress_allow_list();
-        assert_eq!(
-            egress,
-            EgressDomains {
-                exact: vec!["google.com".to_string()],
-                wildcard: vec![".evervault.com".to_string()],
-                allow_all: false
-            }
-        );
-        std::env::remove_var("EV_EGRESS_ALLOW_LIST");
-    }
-
-    fn test_backwards_compat() {
-        std::env::set_var("EV_EGRESS_ALLOW_LIST", "");
-        let egress = get_egress_allow_list();
-        assert_eq!(
-            egress,
-            EgressDomains {
-                exact: vec!["".to_string()],
-                wildcard: vec![],
-                allow_all: true
-            }
-        );
-        std::env::remove_var("EV_EGRESS_ALLOW_LIST")
-    }
-}
