+++ data-plane/src/main.rs
@@ -64,7 +64,7 @@ async fn start(data_plane_port: u16) {
 
     StatsClient::init();
     let ports = configuration::get_egress_ports();
-    let allowed_domains = configuration::get_egress_allow_list();
+    let allowed_domains = shared::server::egress::get_egress_allow_list();
     let egress_proxies = join_all(
         ports
             .into_iter()
