+++ shared/Cargo.toml
@@ -21,6 +21,7 @@ lazy_static = "1.4.0"
 derive_builder = "0.12.0"
 chrono = "0.4.23"
 rand = "^0.8"
+tls-parser = { version = "0.11.0" }
 
 [lib]
 name = "shared"
